refactor: get-user 디렉토리 제거
