+++ src/features/user/api.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query"
-import { fetchUser } from "../../../entities/user/api.ts"
+import { fetchUser } from "../../entities/user/api.ts"
 
 export const useGetUser = (userId: number | null) => {
   return useQuery({
